Data.POrder

Description

This module provides the POrder and related classes

Documentation

class POrder a where

This class indicates that the type a is partially ordered by some relation ⊑.

The class does not actually have a method for ⊑, because we do not need it at runtime. Nevertheless the order better exists for the safety of this API.

This order may be unrelated to the total order given by Ord.

Methods

eqOfLe :: a -> a -> Bool

The eqOfLe method checks _related_ elements for equality.

Formally: For all x ⊑ y, eqOfLe x y == True iff x == y.

This can be more efficient than testing for equality. For example for sets, (==) needs to compare the elements, but eqOfLe only needs to compare sizes. It is always ok to use (==) here.

class Bottom a where

A class indicating that the type a is has a bottom element.

Methods

bottom :: a

class POrder a => Top a where

A class indicating that the type a is has a top element.

Methods

top :: a