Data.Propagator.Class

Description

This module provides the Propagator class

Documentation

class Propagator p x | p -> x where

The Propagator class defines some functions shared by different propagator implementations. This backs the generic Data.Propagator.Purify wrapper.

Methods

newProp :: IO p

newConstProp :: x -> IO p

freezeProp :: p -> IO ()

readProp :: p -> IO x

data WriteToFrozenPropagatorException

Exception thrown by a propagator when attempting to change the value of a frozen propagator

Constructors

WriteToFrozenPropagatorException