Data.Recursive.DualBool

Description

The type RDualBool is like Bool, but allows recursive definitions:

>>> :{
  let x = RDB.true
      y = x RDB.&& z
      z = y RDB.|| RDB.false
  in RDB.get x
:}
True

This finds the greatest solution, i.e. prefers True over False:

>>> :{
  let x = x RDB.&& y
      y = y RDB.&& x
  in (RDB.get x, RDB.get y)
:}
(True,True)

Use RBool from Data.Recursive.Bool if you want the least solution.

Documentation

data RDualBool

Like Bool, but admits recursive definitions, preferring the greatest solution.

get :: RDualBool -> Bool

Extracts the value of a RDualBool

mk :: Bool -> RDualBool

RDB.get (RDB.mk b) === b

true :: RDualBool

RDB.get RDB.true == True

false :: RDualBool

RDB.get RDB.false == False

(&&) :: RDualBool -> RDualBool -> RDualBool

RDB.get (r1 RDB.&& r2) === (RDB.get r1 && RDB.get r2)

(||) :: RDualBool -> RDualBool -> RDualBool

RDB.get (r1 RDB.|| r2) === (RDB.get r1 || RDB.get r2)

and :: [RDualBool] -> RDualBool

RDB.get (RDB.and rs) === and (map RDB.get rs)

or :: [RDualBool] -> RDualBool

RDB.get (RDB.or rs) === or (map RDB.get rs)

not :: RBool -> RDualBool

RDB.get (RDB.not r1) === not (RB.get r1)

id :: RDualBool -> RDualBool

The identity function. This is useful when tying the knot, to avoid a loop that bottoms out:

let x = x in RDB.get x

will not work, but

>>> let x = RDB.id x in RDB.get x
True

does.

| prop> RDB.get (RDB.id r) === RDB.get r